Rename tools to utils
