+++ parcels/particle.py
@@ -10,7 +10,7 @@ import numpy as np
 from parcels._compat import _attrgetter_helper
 from parcels._core.utils.time import TimeInterval
 from parcels._reprs import _format_list_items_multiline
-from parcels.tools.statuscodes import StatusCode
+from parcels.utils.statuscodes import StatusCode
 
 __all__ = ["KernelParticle", "Particle", "ParticleClass", "Variable"]
 _TO_WRITE_OPTIONS = [True, False, "once"]
