+++ parcels/particlefile.py
@@ -15,7 +15,7 @@ from zarr.storage import DirectoryStore
 
 import parcels
 from parcels.particle import _SAME_AS_FIELDSET_TIME_INTERVAL, ParticleClass
-from parcels.tools._helpers import timedelta_to_float
+from parcels.utils._helpers import timedelta_to_float
 
 if TYPE_CHECKING:
     from parcels._core.utils.time import TimeInterval
