+++ parcels/particleset.py
@@ -14,9 +14,9 @@ from parcels._reprs import particleset_repr
 from parcels.kernel import Kernel
 from parcels.kernels import AdvectionRK4
 from parcels.particle import KernelParticle, Particle, create_particle_data
-from parcels.tools.converters import convert_to_flat_array
+from parcels.utils.converters import convert_to_flat_array
 from parcels._logger import logger
-from parcels.tools.statuscodes import StatusCode
+from parcels.utils.statuscodes import StatusCode
 from parcels._warnings import ParticleSetWarning
 
 __all__ = ["ParticleSet"]
