+++ tests-v3/tools/test_helpers.py
@@ -3,8 +3,8 @@ from datetime import timedelta
 import numpy as np
 import pytest
 
-import parcels.tools._helpers as helpers
-from parcels.tools._helpers import deprecated, deprecated_made_private, timedelta_to_float
+import parcels.utils._helpers as helpers
+from parcels.utils._helpers import deprecated, deprecated_made_private, timedelta_to_float
 
 
 def test_format_list_items_multiline():
