+++ tests/common_kernels.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from parcels.tools.statuscodes import StatusCode
+from parcels.utils.statuscodes import StatusCode
 
 
 def DoNothing(particles, fieldset):  # pragma: no cover
