+++ tests/test_interpolation.py
@@ -19,7 +19,7 @@ from parcels.interpolators import (
 from parcels.particle import Particle, Variable
 from parcels.particlefile import ParticleFile
 from parcels.particleset import ParticleSet
-from parcels.tools.statuscodes import StatusCode
+from parcels.utils.statuscodes import StatusCode
 from parcels.uxgrid import UxGrid
 from parcels.xgrid import XGrid
 from tests.utils import TEST_DATA
